requirements.txt
alibabacloud-ocr-api20210707
alibabacloud_darabonba_stream
pdfplumber
pandas
openpyxl
jsonlines
python-dotenv
PyMuPDF
pillow
